refactor(stream): give each converter its own *_converter.go file
Move the four inline converters (ChatToResponses, ResponsesToChat,
AnthropicBetaToResponses, AnthropicToOpenAI) into dedicated
*_converter.go files, matching the existing layout used by the
OpenAI->Anthropic and Responses->Anthropic converters. Pure code move;
no logic changes. Handle* entry points and SSE writers stay in the
original files.
